+++ src/plugins/index.ts
@@ -12,7 +12,6 @@ export * from './store'
 export * from './naive-ui-plugin'
 export * from './dynamic-components'
 export * from './passive-scroll'
-export * from './markdown'
 export * from './analytics'
 export * from './request-core' // 🆕 Request Core 插件
 export * from './layout' // 🆕 布局系统插件